parseFloat to parseInt

+++ src/components/AssetList.js
@@ -13,7 +13,7 @@ const AssetList = ({ assets, handleDeleteAsset }) => {
                 <p className="mb-1">Price: ${asset.market_price ? Number(asset.market_price).toLocaleString("en-US", { minimumFractionDigits: 2, maximumFractionDigits: 2 }) : "N/A"}</p>
                 <p className="mb-1">Quantity: {asset.quantity} shares </p>
                 {/* lots of funny type business going on here with market_price and quantity - dig into it better to see how the backend is sending over these values */}
-                <p className="mb-1">Total Market Value: ${ (parseFloat(asset.market_price) * parseInt(asset.quantity, 10)).toLocaleString('en-US', { minimumFractionDigits: 2, maximumFractionDigits: 2 })}
+                <p className="mb-1">Total Market Value: ${ (parseFloat(asset.market_price) * parseFloat(asset.quantity, 10)).toLocaleString('en-US', { minimumFractionDigits: 2, maximumFractionDigits: 2 })}
                 </p>
               </div>
               <button
